requirements.txt
mcp>=0.1.0
pandas>=2.0.0
google-auth>=2.23.0
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0